requirements.txt
asgiref==3.7.2
Django==4.2.7
django-admin==2.0.2
django-excel-response2==3.0.6
django-six==1.0.5
excel-base==1.0.4
isoweek==1.3.3
python-dateutil==2.8.2
python-dotenv==1.0.0
screen==1.0.1
six==1.16.0
sqlparse==0.4.4
TimeConvert==3.0.13
tzlocal==5.2
xlwt==1.3.0

djangorestframework~=3.14.0